Mono: More natural NPM scripts feel

+++ bin/tasks/test.js
@@ -2,12 +2,12 @@ const { spawn } = require('child_process')
 const { execSync } = require('child_process')
 const onStart = require('../helpers/on-start')
 
-async function test(PROJECT_NAME) {
+function test(projectName, kill = true) {
   const devProcess = spawn('npm start', [], { shell: true, detached: true })
 
   onStart(function() {
     try {
-      const testPattern = `${__dirname}/../src/projects/${PROJECT_NAME}/tests/*.test.js`
+      const testPattern = `${__dirname}/../src/projects/${projectName}/tests/*.test.js`
       execSync(`mocha -p tsconfig.json ${testPattern}`, { stdio: 'inherit' })
     }
 
@@ -16,9 +16,22 @@ async function test(PROJECT_NAME) {
       process.exit()
     }
 
-    process.kill(-devProcess.pid)
-    process.exit()
+    if (kill) {
+      process.kill(-devProcess.pid)
+      process.exit()
+    }
   })
 }
 
-module.exports = test
+const testAll = function() {
+  const projects = readdirSync(join(__dirname, '..', '..', 'projects'))
+
+  for (let p = 0; p < projects.length; p ++) {
+    const projectName = projects[p]
+    if (statSync(join(__dirname, '..', '..', 'projects', projectName)).isDirectory()) {
+      test(projectName, false)
+    }
+  }
+}
+
+module.exports = { test, testAll }
